feat(plugins/auth): move user functions to seperate folder
