+++ plugins/auth/methods/index.ts
@@ -1,0 +1,2 @@
+export { default as user } from './user';
+export { default as credentials } from './credentials';
