+++ plugins/auth/methods/user/accept.ts
