+++ plugins/auth/methods/user/modify.ts
