+++ plugins/auth/methods/user/remove.ts
