+++ plugins/auth/methods/user/scope.ts
